+++ Localization/deDE.lua
@@ -110,6 +110,7 @@ L.Settings.IndicateInterruptsTooltip =
 	"Desaturiert das Icon, zeigt einen Indikator an und verzögert das Ausblenden des Icons um eine Sekunde. Funktioniert nicht bei kanalisierten Zaubern."
 
 L.Settings.ShowSwipeLabel = "Abklingzeitsanimation anzeigen"
+L.Settings.ShowSwipeTooltip = nil
 
 L.Settings.ShowBorderLabel = "Rahmen"
 L.Settings.ShowBorderTooltip = nil